fix(release): gradle releaseVersion should not tag
jgit out of the box doesn't support ssh commit signing which
means that releaseVersion fails because it can't parse the
`gpg.format=ssh` part of the config.
Force releaseVersion to run in dry-run mode by not allowing
it to tag; and then grep the version number it thinks we
should.
Finally tag using vanilla git.